+++ core/src/main/java/org/nervos/ckb/transaction/TransactionInput.java
@@ -7,10 +7,10 @@ import org.nervos.ckb.type.cell.CellInput;
 public class TransactionInput {
 
   public CellInput input;
-  public BigInteger capacity;
+  public long capacity;
   public String lockHash;
 
-  public TransactionInput(CellInput input, BigInteger capacity, String lockHash) {
+  public TransactionInput(CellInput input, long capacity, String lockHash) {
     this.input = input;
     this.capacity = capacity;
     this.lockHash = lockHash;
